Symmetry adapted pool constructed at c++

+++ src/qforte/find_irrep.cc
@@ -1,10 +1,10 @@
 #include "find_irrep.h"
 
-int find_irrep(const std::vector<int>& orb_irreps_to_int,
-        const std::vector<int>& spinorb_indices) {
+size_t find_irrep(const std::vector<size_t>& orb_irreps_to_int,
+        const std::vector<size_t>& spinorb_indices) {
 
-    int irrep = 0;
-    for (int index : spinorb_indices) {
+    size_t irrep = 0;
+    for (size_t index : spinorb_indices) {
         irrep ^= orb_irreps_to_int[index/2];
     }
